Refactor AppManager, use separate Initializers
